First version of MuonProcessor
